backend/requirements.txt
asgiref==3.3.4
CouchDB==1.2
Django==3.2
django-cors-headers==3.7.0
django-geojson==3.2.0
django-webpack-loader==0.7.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
geographiclib==1.50
geopy==2.1.0
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
ujson==4.0.2
